feat: configure Drupal file paths in Nginx Dockerfile (#11)
This change sets up the default environment variables and file paths for a Drupal
application running in an Nginx container. It creates a symlink to the default
location where the persistent files are stored, which is usually done at runtime
by calling the makefiles script. This change ensures that the file paths are
properly configured without having to run the script every time the container
is started.