feat(drupal-tailscale): Add script to save environment variables
This change adds a script that saves a predefined list of environment
variables to the `/etc/profile` file. This is a temporary solution to
handle a bug/limitation in the way Tailscale SSH runs commands, as
described in the issue https://github.com/tailscale/tailscale/issues/12395.
Until the bug is fixed, this script ensures that the environment
variables are available when the shell is started by Tailscale SSH.
feat(env): remove quotes from environment variable exports

The changes in this commit remove the quotes around the environment variable
exports in the `save-env.sh` script. This ensures that the exported variables
can be properly consumed by other scripts or processes, as the quotes are not
necessary and can cause issues in certain scenarios.

The key changes are:

- Removing the single quotes around the `$env_var` and `$var_name` variables
  when echoing them to the destination file.
- Renaming the `env_vars` array to `var_names` to better reflect the purpose of
  the array.
feat: update environment file location and logic

Modify the `save-env.sh` script to allow for a custom SSH directory,
and update the destination file path to use the `environment` file
in the SSH directory instead of the `.profile` file in the user's
home directory. This provides more flexibility in where the
environment variables are stored.
feat: Persist environment variables in multiple shell profiles

Ensures that the environment variables are loaded by any shell by
copying the `.shrc` file to both the `.bashrc` and `.bash_profile`
files. This helps to ensure that the environment variables are
available in any shell session.